Share VarRenamer code.
