+++ crates/cairo-lang-lowering/src/optimizations/mod.rs
@@ -9,3 +9,4 @@ pub mod return_optimization;
 pub mod scrub_units;
 pub mod split_structs;
 pub mod strategy;
+pub mod var_renamer;
